+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- `grant update` command for self-updating the binary via GitHub Releases using `rhysd/go-github-selfupdate`
+
 ## [0.3.0] - 2026-02-19
 
 ### Added